remove GCC warning
git-svn-id: svn://svn.code.sf.net/p/irrlicht/code@292 dfc29bdd-3216-0410-991c-e03cc46cb475

+++ source/Irrlicht/CImageWriterTGA.cpp
@@ -124,7 +124,7 @@ bool CImageWriterTGA::writeImage(io::IWriteFile *file, IImage *image)
 	imageFooter.DeveloperOffset = 0;
 	strncpy(imageFooter.Signature, "TRUEVISION-XFILE.", 18);
 
-	if (file->write(&imageFooter, sizeof(imageFooter)) < sizeof(imageFooter))
+	if (file->write(&imageFooter, sizeof(imageFooter)) < (s32)sizeof(imageFooter))
 		return false;
 
 	return imageHeader.ImageHeight < y;
